+++ src/main/java/com/example/demo/DTO/UtilisateurResponseDto.java
@@ -30,7 +30,7 @@ public class UtilisateurResponseDto {
 	private String poste;
 	private String email;
 	private String adresse;
-	private Integer telephone;	
+	private long telephone;	
 
 	private Leservice leservice;
